fix(lint): split routines/command.rs under the 500-line gate
Moves the PATH/tmux/agent-binary resolution helpers (bin_dir,
tmux_available, agent_command_available, resolve_tmux_bin, cron_path,
and friends) into a new sibling command_path_resolution.rs, re-exported
via `pub(crate) use`. Pure code move, no behavior change.
